pyproject.toml
[project]
name = "pred-genai"
version = "0.1.0"
description = ""
authors = [
    {name = "vadimglinskiy"}
]
readme = "README.md"
requires-python = ">=3.12"
package-mode = false
dependencies = [
    "absl-py (>=2.2.2,<3.0.0)",
    "aiohappyeyeballs (>=2.6.1,<3.0.0)",
    "aiohttp (>=3.11.18,<4.0.0)",
    "aiosignal (>=1.3.2,<2.0.0)",
    "annotated-types (>=0.7.0,<0.8.0)",
    "anyio (>=4.9.0,<5.0.0)",
    "argcomplete (>=3.6.2,<4.0.0)",
    "channels-redis (>=4.2.0,<5.0.0)",
    "asgiref (>=3.0,<4.0)",
    "astunparse (>=1.6.3,<2.0.0)",
    "attrs (>=25.3.0,<26.0.0)",
    "autobahn (>=24.4.2,<25.0.0)",
    "automat (>=25.4.16,<26.0.0)",
    "azure-core (>=1.33.0,<2.0.0)",
    "azure-identity (>=1.21.0,<2.0.0)",
    "av (>=14.3.0,<15.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "build (>=1.2.2.post1,<2.0.0)",
    "cachetools (>=5.5.2,<6.0.0)",
    "certifi (>=2025.4.26,<2026.0.0)",
    "cffi (>=1.17.1,<2.0.0)",
    "chardet (>=5.2.0,<6.0.0)",
    "charset-normalizer (>=3.4.1,<4.0.0)",
    "chroma-hnswlib (>=0.7.3,<0.8.0)",
    "chromadb (>=0.4.22,<0.5.0)",
    "click (>=8.1.8,<9.0.0)",
    "cobble (>=0.1.4,<0.2.0)",
    "coloredlogs (>=15.0.1,<16.0.0)",
    "compressed-rtf (>=1.0.7,<2.0.0)",
    "constantly (>=23.10.4,<24.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "deprecated (>=1.2.18,<2.0.0)",
    "distro (>=1.9.0,<2.0.0)",
    "django-channels (>=0.7.0,<0.8.0)",
    "dnspython (>=2.7.0,<3.0.0)",
    "docx2txt (>=0.9,<0.10)",
    "durationpy (>=0.9,<0.10)",
    "ebcdic (>=1.1.1,<2.0.0)",
    "extract-msg (>=0.54.1,<0.55.0)",
    "fastapi (==0.115.9)",
    "filelock (>=3.18.0,<4.0.0)",
    "flatbuffers (>=25.2.10,<26.0.0)",
    "frozenlist (>=1.6.0,<2.0.0)",
    "fsspec (>=2025.3.2,<2026.0.0)",
    "gast (>=0.6.0,<0.7.0)",
    "google-auth (>=2.39.0,<3.0.0)",
    "google-pasta (>=0.2.0,<0.3.0)",
    "googleapis-common-protos (>=1.70.0,<2.0.0)",
    "grpcio (>=1.71.0,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "h11 (>=0.16.0,<0.17.0)",
    "h2 (>=4.2.0,<5.0.0)",
    "h5py (>=3.13.0,<4.0.0)",
    "hpack (>=4.1.0,<5.0.0)",
    "httpcore (>=1.0.9,<2.0.0)",
    "httptools (>=0.6.4,<0.7.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "huggingface-hub (>=0.30.2,<0.31.0)",
    "humanfriendly (>=10.0,<11.0)",
    "hyperframe (>=6.1.0,<7.0.0)",
    "hyperlink (>=21.0.0,<22.0.0)",
    "idna (>=3.10,<4.0)",
    "imapclient (>=3.0.1,<4.0.0)",
    "importlib-metadata (>=6.0.0,<7.0.0)",
    "importlib-resources (>=6.5.2,<7.0.0)",
    "incremental (>=24.7.2,<25.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "jiter (>=0.9.0,<0.10.0)",
    "joblib (>=1.4.2,<2.0.0)",
    "keras (>=3.9.2,<4.0.0)",
    "kubernetes (>=32.0.1,<33.0.0)",
    "libclang (>=18.1.1,<19.0.0)",
    "lxml (>=5.4.0,<6.0.0)",
    "mammoth (>=1.9.0,<2.0.0)",
    "markdown (>=3.8,<4.0)",
    "markdown-it-py (>=3.0.0,<4.0.0)",
    "markupsafe (>=3.0.2,<4.0.0)",
    "mdurl (>=0.1.2,<0.2.0)",
    "ml-dtypes (>=0.3.1,<0.4.0)",
    "mmh3 (>=5.1.0,<6.0.0)",
    "monotonic (>=1.6,<2.0)",
    "mpmath (>=1.3.0,<2.0.0)",
    "msal (>=1.32.3,<2.0.0)",
    "msal-extensions (>=1.3.1,<2.0.0)",
    "msgpack-python (>=0.5.6,<0.6.0)",
    "multidict (>=6.4.3,<7.0.0)",
    "namex (>=0.0.9,<0.0.10)",
    "networkx (>=3.4.2,<4.0.0)",
    "nslookup (>=1.8.1,<2.0.0)",
    "numpy (>=1.24.0,<2.0.0)",
    "oauthlib (>=3.2.2,<4.0.0)",
    "olefile (>=0.47,<0.48)",
    "onnxruntime (>=1.21.1,<2.0.0)",
    "openai (>=1.76.0,<2.0.0)",
    "opt-einsum (>=3.4.0,<4.0.0)",
    "optree (>=0.15.0,<0.16.0)",
    "orjson (>=3.10.16,<4.0.0)",
    "overrides (>=7.7.0,<8.0.0)",
    "packaging (>=25.0,<26.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pdfminer-six (>=20250416,<20250417)",
    "pillow (>=11.2.1,<12.0.0)",
    "pip-chill (>=1.0.3,<2.0.0)",
    "portalocker (>=3.1.1,<4.0.0)",
    "posthog (>=4.0.0,<5.0.0)",
    "propcache (>=0.3.1,<0.4.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "pyasn1 (>=0.6.1,<0.7.0)",
    "pyasn1-modules (>=0.4.2,<0.5.0)",
    "pycparser (>=2.22,<3.0)",
    "pycryptodome (>=3.22.0,<4.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "pygments (>=2.19.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pyopenssl (>=25.0.0,<26.0.0)",
    "pypika (>=0.48.9,<0.49.0)",
    "pyproject-hooks (>=1.2.0,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "regex (>=2024.11.6,<2025.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "requests-oauthlib (>=2.0.0,<3.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "safetensors (>=0.5.3,<0.6.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "scipy (>=1.15.2,<2.0.0)",
    "service-identity (>=24.2.0,<25.0.0)",
    "setuptools (>=80.0.0,<81.0.0)",
    "shellingham (>=1.5.4,<2.0.0)",
    "six (>=1.17.0,<2.0.0)",
    "sniffio (>=1.3.1,<2.0.0)",
    "sortedcontainers (>=2.4.0,<3.0.0)",
    "soupsieve (>=2.7,<3.0)",
    "speechrecognition (>=3.14.2,<4.0.0)",
    "sqlparse (>=0.5.3,<0.6.0)",
    "starlette (==0.40.0)",
    "sympy (>=1.14.0,<2.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "tensorboard (>=2.19.0,<3.0.0)",
    "tensorboard-data-server (>=0.7.2,<0.8.0)",
    "termcolor (>=3.0.1,<4.0.0)",
    "threadpoolctl (>=3.6.0,<4.0.0)",
    "time-machine (>=2.16.0,<3.0.0)",
    "tokenizers (>=0.21.1,<0.22.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "twisted (>=24.11.0,<25.0.0)",
    "txaio (>=23.1.1,<24.0.0)",
    "typer (>=0.15.2,<0.16.0)",
    "typing-extensions (>=4.13.2,<5.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "tzlocal (>=5.3.1,<6.0.0)",
    "urllib3 (>=2.4.0,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "uvloop (>=0.21.0,<0.22.0)",
    "watchfiles (>=1.0.5,<2.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "wheel (>=0.45.1,<0.46.0)",
    "wrapt (>=1.17.2,<2.0.0)",
    "xlrd (>=2.0.1,<3.0.0)",
    "xlsxwriter (>=3.2.3,<4.0.0)",
    "yarl (>=1.20.0,<2.0.0)",
    "zipp (>=3.21.0,<4.0.0)",
    "zope-interface (>=7.2,<8.0)",
    "llama-cpp-python (>=0.3.8,<0.4.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "aiortc (>=1.11.0,<2.0.0)",
    "pyee (>=13.0.0,<14.0.0)",
    "pylibsrtp (>=0.12.0,<0.13.0)",
    "aioice (>=0.10.1,<0.11.0)",
    "ifaddr (>=0.2.0,<0.3.0)",
    "google-crc32c (>=1.7.1,<2.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "jsonschema (>=4.23.0,<5.0.0)",
    "referencing (>=0.36.2,<0.37.0)",
    "rpds (>=5.1.0,<6.0.0)",
    "jsonschema-specifications (>=2025.4.1,<2026.0.0)",
    "ffmpeg (>=1.4,<2.0)",
    "chromadb (>=0.4.22,<0.5.0)",
    "opentelemetry-api (>=1.22.0,<2.0.0)",
    "opentelemetry-sdk (>=1.22.0,<2.0.0)",
    "opentelemetry-exporter-otlp-proto-grpc (>=1.22.0,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.41b0,<0.54)",
    "opentelemetry-instrumentation-asgi (>=0.41b0,<0.54)",
    "djangorestframework (>=3.15.0,<4.0.0)",
    "transformers (>=4.38.0,<5.0.0)",
    "torch (>=2.2.0,<3.0.0)",
    "pgvector (>=0.2.4)",
    "psycopg2-binary (>=2.9.9)",
    "sqlalchemy (>=2.0.28,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

